pyproject.toml
[project]
name = "mcp-gsuite"
version = "0.4.0"
description = "MCP Server to connect to Google G-Suite"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "beautifulsoup4>=4.12.3",
 "google-api-python-client>=2.154.0",
 "httplib2>=0.22.0",
 "mcp>=1.3.0",
 "google-auth-oauthlib>=1.0.0", # Added for new auth flow
 "google-auth-library-python>=2.0.0", # Added for core auth functionality
 "python-dotenv>=1.0.1",
 "pytz>=2024.2",
 "requests>=2.32.3",
 "redis>=5.0"
]

[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
    "pytest>=8.0", # Added pytest for running tests
]

[project.scripts]
mcp-gsuite = "mcp_gsuite:main"
